+++ dune/grid/common/defaultgridview.hh
@@ -37,8 +37,12 @@ namespace Dune
     typedef typename Grid :: Traits :: LevelIntersectionIterator
     IntersectionIterator;
 
-    /** \brief type of the collective communication */
-    typedef typename Grid :: Traits :: CollectiveCommunication CollectiveCommunication;
+    /** \brief type of the communication */
+    typedef typename Grid :: Traits :: Communication Communication;
+
+    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
+    [[deprecated("Use Communication instead!")]]
+    typedef Communication CollectiveCommunication;
 
     template< int cd >
     struct Codim
@@ -88,8 +92,12 @@ namespace Dune
     /** \brief type of the intersection iterator */
     typedef typename Traits :: IntersectionIterator IntersectionIterator;
 
-    /** \brief type of the collective communication */
-    typedef typename Traits :: CollectiveCommunication CollectiveCommunication;
+    /** \brief type of the communication */
+    typedef typename Grid :: Traits :: Communication Communication;
+
+    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
+    [[deprecated("Use Communication instead!")]]
+    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >
@@ -172,8 +180,8 @@ namespace Dune
       return entity.impl().ilevelend();
     }
 
-    /** \brief obtain collective communication object */
-    const CollectiveCommunication &comm () const
+    /** \brief obtain communication object */
+    const Communication &comm () const
     {
       return grid().comm();
     }
@@ -222,8 +230,12 @@ namespace Dune
     typedef typename Grid :: Traits :: LeafIntersectionIterator
     IntersectionIterator;
 
-    /** \brief type of the collective communication */
-    typedef typename Grid :: Traits :: CollectiveCommunication CollectiveCommunication;
+    /** \brief type of the communication */
+    typedef typename Grid :: Traits :: Communication Communication;
+
+    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
+    [[deprecated("Use Communication instead!")]]
+    typedef Communication CollectiveCommunication;
 
     template< int cd >
     struct Codim
@@ -273,8 +285,12 @@ namespace Dune
     /** \brief type of the intersection iterator */
     typedef typename Traits :: IntersectionIterator IntersectionIterator;
 
-    /** \brief type of the collective communication */
-    typedef typename Traits :: CollectiveCommunication CollectiveCommunication;
+    /** \brief type of the communication */
+    typedef typename Grid :: Traits :: Communication Communication;
+
+    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
+    [[deprecated("Use Communication instead!")]]
+    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >
@@ -357,8 +373,8 @@ namespace Dune
       return entity.impl().ileafend();
     }
 
-    /** \brief obtain collective communication object */
-    const CollectiveCommunication &comm () const
+    /** \brief obtain communication object */
+    const Communication &comm () const
     {
       return grid().comm();
     }