+++ dune/grid/geometrygrid/gridfamily.hh
@@ -106,7 +106,13 @@ namespace Dune
         typedef GeoGrid::IdSet< const Grid, typename HostGrid::Traits::LocalIdSet >
         LocalIdSet;
 
-        typedef typename HostGrid::Traits::CollectiveCommunication CollectiveCommunication;
+        typedef typename HostGrid::Traits::Communication Communication;
+
+        /**
+         * \deprecated Use Communication instead! Will be removed after Dune 2.9.
+         */
+        [[deprecated("Use Communication instead!")]]
+        typedef Communication CollectiveCommunication;
 
         typedef Dune::GridView< GeoGrid::GridViewTraits< typename HostGrid::LeafGridView, CoordFunction, Allocator > > LeafGridView;
         typedef Dune::GridView< GeoGrid::GridViewTraits< typename HostGrid::LevelGridView, CoordFunction, Allocator > > LevelGridView;