+++ dune/grid/io/file/vtk/volumeiterators.hh
@@ -42,7 +42,7 @@ namespace Dune {
 
       typedef ConformingConnectivityWriter<ConformingVolumeIteratorFactory<GV>
           > ConnectivityWriter;
-      typedef typename GV::CollectiveCommunication CollectiveCommunication;
+      typedef typename GV::Communication Communication;
 
       explicit ConformingVolumeIteratorFactory(const GV& gv_)
         : gv(gv_)
@@ -72,7 +72,7 @@ namespace Dune {
       ConnectivityWriter makeConnectivity() const {
         return ConnectivityWriter(*this);
       }
-      const CollectiveCommunication& comm() const {
+      const Communication& comm() const {
         return gv.comm();
       }
     };
@@ -95,7 +95,7 @@ namespace Dune {
       typedef CornerIterator PointIterator;
 
       typedef NonConformingConnectivityWriter<Cell> ConnectivityWriter;
-      typedef typename GV::CollectiveCommunication CollectiveCommunication;
+      typedef typename GV::Communication Communication;
 
       explicit NonConformingVolumeIteratorFactory(const GV& gv_)
         : gv(gv_)
@@ -121,7 +121,7 @@ namespace Dune {
       ConnectivityWriter makeConnectivity() const {
         return ConnectivityWriter();
       }
-      const CollectiveCommunication& comm() const {
+      const Communication& comm() const {
         return gv.comm();
       }
     };